adds new file under utils folder to hold teamMembers data
