+++ src/components/about-us-helpers/Leadership.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-
+import {teamMembers} from '../../utils/teamMembers.js'
 import {
   bgSquare1,
   bgSquare2,
@@ -8,69 +8,6 @@ import {
   section2,
   card,
 } from './Leadership.module.scss'
-
-const teamMembers = [
-  {
-    name: 'Skyler Dowdy',
-    role: 'Team Lead',
-    imagePath: require(`../../assets/images/team-members/skyler.jpg`),
-  },
-  {
-    name: 'Jan Patrick Eliares',
-    role: 'UX/UI Designer',
-    imagePath: require(`../../assets/images/team-members/jp.png`),
-  },
-  {
-    name: 'Nora Barazanchi',
-    role: 'UX/UI Designer',
-    imagePath: ``,
-  },
-  {
-    name: 'Louis Gelinas',
-    role: 'Engineer',
-    imagePath: require(`../../assets/images/team-members/louis.jpg`),
-  },
-  {
-    name: 'Mark King',
-    role: 'Engineer',
-    imagePath: require(`../../assets/images/team-members/mark.jpg`),
-  },
-  {
-    name: 'Westley Strellis',
-    role: 'Engineer',
-    imagePath: require(`../../assets/images/team-members/westley.png`),
-  },
-  {
-    name: 'Basil Havens',
-    role: 'Mobile Engineer',
-    imagePath: require(`../../assets/images/team-members/basil.jpg`),
-  },
-  {
-    name: 'Ben Rogers',
-    role: 'Engineer',
-    imagePath: require(`../../assets/images/team-members/ben.jpg`),
-  },
-  {
-    name: 'Tyler Berrett',
-    role: 'Mobile Engineer',
-    imagePath: require(`../../assets/images/team-members/tyler.jpg`),
-  },
-  {
-    name: 'Jordan Christensen',
-    role: 'Mobile Engineer',
-    imagePath: require(`../../assets/images/team-members/jordan.png`),
-  },
-  {
-    name: 'Justin Gent',
-    role: 'Mobile Engineer',
-    imagePath: require(`../../assets/images/team-members/justin.jpg`),
-  },
-  {
-    name: 'Lowell Jacobs',
-    role: 'Engineer',
-    imagePath: require(`../../assets/images/team-members/lowell.jpg`),
-  },  
-]
 
 const Leadership = () => {
   return (
