+++ src/utils/teamMembers.js
@@ -1,0 +1,64 @@
+const teamMembers = [
+  {
+    name: 'Skyler Dowdy',
+    role: 'Team Lead',
+    imagePath: require(`../assets/images/team-members/skyler.jpg`),
+  },
+  {
+    name: 'Jan Patrick Eliares',
+    role: 'UX/UI Designer',
+    imagePath: require(`../assets/images/team-members/jp.png`),
+  },
+  {
+    name: 'Nora Barazanchi',
+    role: 'UX/UI Designer',
+    imagePath: ``,
+  },
+  {
+    name: 'Louis Gelinas',
+    role: 'Engineer',
+    imagePath: require(`../assets/images/team-members/louis.jpg`),
+  },
+  {
+    name: 'Mark King',
+    role: 'Engineer',
+    imagePath: require(`../assets/images/team-members/mark.jpg`),
+  },
+  {
+    name: 'Westley Strellis',
+    role: 'Engineer',
+    imagePath: require(`../assets/images/team-members/westley.png`),
+  },
+  {
+    name: 'Basil Havens',
+    role: 'Mobile Engineer',
+    imagePath: require(`../assets/images/team-members/basil.jpg`),
+  },
+  {
+    name: 'Ben Rogers',
+    role: 'Engineer',
+    imagePath: require(`../assets/images/team-members/ben.jpg`),
+  },
+  {
+    name: 'Tyler Berrett',
+    role: 'Mobile Engineer',
+    imagePath: require(`../assets/images/team-members/tyler.jpg`),
+  },
+  {
+    name: 'Jordan Christensen',
+    role: 'Mobile Engineer',
+    imagePath: require(`../assets/images/team-members/jordan.png`),
+  },
+  {
+    name: 'Justin Gent',
+    role: 'Mobile Engineer',
+    imagePath: require(`../assets/images/team-members/justin.jpg`),
+  },
+  {
+    name: 'Lowell Jacobs',
+    role: 'Engineer',
+    imagePath: require(`../assets/images/team-members/lowell.jpg`),
+  },
+]
+
+export {teamMembers}
